re-use entangler_map helpers and clean up the codes
