+++ qiskit_aqua/utils/__init__.py
@@ -21,6 +21,7 @@ from .jsonutils import convert_dict_to_json,convert_json_to_dict
 from .random_matrix_generator import random_unitary, random_h2_body, random_h1_body
 from .decimal_to_binary import decimal_to_binary
 from .summarize_circuits import summarize_circuits
+from .entangler_map import get_entangler_map, validate_entangler_map
 
 __all__ = ['tensorproduct',
            'PauliGraph',
@@ -30,4 +31,6 @@ __all__ = ['tensorproduct',
            'random_h2_body',
            'random_h1_body',
            'decimal_to_binary',
-           'summarize_circuits']
+           'summarize_circuits',
+           'get_entangler_map',
+           'validate_entangler_map']
