+++ qiskit_aqua/utils/feature_extractions/second_order_expansion.py
@@ -69,10 +69,6 @@ class SecondOrderExpansion(FeatureExtraction):
         super().__init__(configuration or self.SECOND_ORDER_EXPANSION_CONFIGURATION.copy())
         self._ret = {}
 
-    def init_params(self, params):
-        args = {k: v for k, v in params.items() if k != 'name'}
-        self.init_args(**args)
-
     def init_args(self, num_qubits, depth, entangler_map=None, entanglement='full'):
         self._num_qubits = num_qubits
         self._depth = depth
